+++ src/views/sidebar/sync-data-handler.tsx
@@ -7,7 +7,7 @@ import React, { FC, useEffect } from 'react';
 
 import { updatePrimaryBadge } from '@zextras/carbonio-shell-ui';
 
-import { useSyncDataHandler } from './commons/sync-data-handler-hooks';
+import { useSyncDataHandler } from './commons/use-sync-data-handler';
 import { FOLDERS } from '../../carbonio-ui-commons/constants/folders';
 import { useFolder } from '../../carbonio-ui-commons/store/zustand/folder';
 import { MAILS_ROUTE } from '../../constants';